% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getviafid.R
\name{getviafid_lccn}
\alias{getviafid_lccn}
\alias{getviafid_sourceids}
\title{Get VIAF cluster IDs}
\usage{
getviafid_lccn(lccns)

getviafid_sourceids(sids)
}
\arguments{
\item{lccns}{A list of Library of Congress Control Numbers}

\item{cache}{Directory for (usually temporary) downloaded responses}

\item{sids}{A list of source IDs in the format ```{SRC}|{number}```, for example,
```DNB|1034425390```.}
}
\value{
A tibble of LC control numbers and corresponding VIAF IDs.

A tibble of source ID numbers and corresponding VIAF IDs.
}
\description{
Get VIAF cluster from LC and other IDs

Get VIAF cluster from LC and other IDs
}
\details{
The [Authority Cluster Resource](https://www.oclc.org/developer/api/oclc-apis/viaf/authority-cluster.en.html)
of the VIAF API provides redirect responses to queries of Library of Congress control numbers
and Source ID's. When queried using the [curl::multi_download()] function of
[Jeroen Ooms's R package curl](https://github.com/jeroen/curl),
one gets a data.frame that contains the final url (after redirects) of the request.
This final url is to the VIAF cluster to which the control number belongs. This url
contains the desired VIAF ID.
}
