% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rviaf.R
\docType{package}
\name{rviaf}
\alias{rviaf}
\alias{rviaf-package}
\title{R Interface to the VIAF (Virtual International Authority File) API}
\description{
The is a fork of Stefanie Schneider's original viafr package. This fork
adds functionality for accessing VIAF data using XSL transform methods.
}
\details{
rviaf
}
\seealso{
\link{cluster_data}
}
\author{
\strong{Maintainer}: Stefanie Schneider \email{stefanie.schneider@itg.uni-muenchen.de} (\href{https://orcid.org/0000-0003-4915-6949}{ORCID})

Authors:
\itemize{
  \item Alan Engel \email{alan@ttcrossroads.net} (\href{https://orcid.org/0000-0003-0070-2347}{ORCID})
}

}
