% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{viaf_normalize}
\alias{viaf_normalize}
\title{Normalization for compatability with VIAF contributors}
\usage{
viaf_normalize(string, source = "ALL")
}
\description{
'viaf_normalize()'' transforms xml nodes imported from viaf.xml, and
marc21.xml files into normalized name strings. The motivating example
is this xml element found in
[VAIF](https://viaf.org/viaf/253482919/viaf.xml)

\preformatted{
<ns1:x400>
<ns1:datafield dtype="MARC21" ind1="1" ind2=" " tag="400">
<ns1:subfield code="a">千葉, 玄彌</ns1:subfield>
<ns1:normalized>仟叶 伭弥</ns1:normalized>
</ns1:datafield>
<ns1:sources>
<ns1:s>NII</ns1:s>
<ns1:sid>NII|DA0067554X</ns1:sid>
</ns1:sources>
</ns1:x400>
}
}
